+++ OfflineDocumentation/BurstTraceReadme.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+namespace Elfinik.BurstTrace.ReadMe
+{
+    public class BurstTraceReadme : ScriptableObject
+    {
+    }
+}